ManualMemory

Documentation for ManualMemory.

ManualMemory.LazyPreserve

ManualMemory.PseudoPtr

ManualMemory.preserve

ManualMemory.preserve_buffer

ManualMemory.LazyPreserve — Type

LazyPreserve(x, ptrcall=nothing)

Used by preserve to identify arguments that will be unwrapped with preserve_buffer, which is in turn converted to a pointer. If ptrcall is specified conversion to a pointer occurs through a call equivalent to ptrcall(preserve_buffer(x), x). Otherwise, a call equivalent to pointer(preserve_buffer(x)) occurs.

ManualMemory.PseudoPtr — Type

PseudoPtr(data, position=firstindex(data))

Provides a convenient wrapper that functions like pointer(data) for instances where data cannot produce a viable pointer.

ManualMemory.preserve — Method

preserve(op, args...; kwargs...)

Searches through args and kwargs for instances of LazyPreserve, which are unwrapped using preserve_buffer and preserved from garbage collection (GC.@preserve). the resulting buffers are converted to pointers and passed in order to op.

Examples

julia> using ManualMemory: store!, preserve, LazyPreserve

julia> x = [0 0; 0 0];

julia> preserve(store!, LazyPreserve(x), 1)

julia> x[1]
1

ManualMemory.preserve_buffer — Method

preserve_buffer(x)

For structs wrapping arrays, using GC.@preserve can trigger heap allocations. preserve_buffer attempts to extract the heap-allocated part. Isolating it by itself will often allow the heap allocations to be elided. For example:

julia> using StaticArrays, BenchmarkTools
julia> # Needed until a release is made featuring https://github.com/JuliaArrays/StaticArrays.jl/commit/a0179213b741c0feebd2fc6a1101a7358a90caed
       Base.elsize(::Type{<:MArray{S,T}}) where {S,T} = sizeof(T)
julia> @noinline foo(A) = unsafe_load(A,1)
foo (generic function with 1 method)
julia> function alloc_test_1()
           A = view(MMatrix{8,8,Float64}(undef), 2:5, 3:7)
           A[begin] = 4
           GC.@preserve A foo(pointer(A))
       end
alloc_test_1 (generic function with 1 method)
julia> function alloc_test_2()
           A = view(MMatrix{8,8,Float64}(undef), 2:5, 3:7)
           A[begin] = 4
           pb = parent(A) # or `LoopVectorization.preserve_buffer(A)`; `perserve_buffer(::SubArray)` calls `parent`
           GC.@preserve pb foo(pointer(A))
       end
alloc_test_2 (generic function with 1 method)
julia> @benchmark alloc_test_1()
BenchmarkTools.Trial:
  memory estimate:  544 bytes
  allocs estimate:  1
  --------------
  minimum time:     17.227 ns (0.00% GC)
  median time:      21.352 ns (0.00% GC)
  mean time:        26.151 ns (13.33% GC)
  maximum time:     571.130 ns (78.53% GC)
  --------------
  samples:          10000
  evals/sample:     998
julia> @benchmark alloc_test_2()
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     3.275 ns (0.00% GC)
  median time:      3.493 ns (0.00% GC)
  mean time:        3.491 ns (0.00% GC)
  maximum time:     4.998 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1000